# Documentation/openapi.yaml
openapi: 3.0.0
info:
  title: Supchat
  description: une API de chat pour une entreprise
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Serveur local Symfony
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Users:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
          minLength: 3
          maxLength: 255
        lastName:
          type: string
          minLength: 3
          maxLength: 255
        userName:
          type: string
          minLength: 3
          maxLength: 255
        emailAddress:
          type: string
          format: email
          minLength: 3
          maxLength: 255
        password:
          type: string
          minLength: 8
          maxLength: 255
          nullable: true
        role:
          type: string
          enum: ["ROLE_USER", "ROLE_ADMIN"]
        theme:
          type: boolean
        status:
          type: string
          minLength: 3
          maxLength: 255
      required: [firstName, lastName, userName, emailAddress, role, theme, status]
...
    Workspaces:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          minLength: 3
          maxLength: 255
        status:
          type: boolean
          description: true = public, false = private
        creator:
          $ref: '#/components/schemas/Users'
      required: [name, status, creator]
...
paths:
  /api/admin/workspaces:
    get:
      tags: [Admin - Workspaces]
      summary: Liste tous les workspaces
      operationId: listAllWorkspaces
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Liste des workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workspaces'
        '401':
          description: Non autorisé (JWT requis)
  /api/admin/workspaces/{id}:
      delete:
        tags: [Admin - Workspaces]
        summary: Désactive un workspace
        operationId: softDeleteWorkspace
        security:
          - BearerAuth: []
        parameters:
          - name: id
            in: path
            required: true
            schema:
              type: integer
        responses:
          '200':
            description: Workspace désactivé avec succès
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
          '404':
            description: Workspace non trouvé
          '401':
            description: Non autorisé

      put:
        tags: [Admin - Workspaces]
        summary: Met à jour un workspace
        operationId: updateWorkspace
        security:
          - BearerAuth: []
        parameters:
          - name: id
            in: path
            required: true
            schema:
              type: integer
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  status:
                    type: boolean
                example:
                  name: "Nouveau nom"
                  status: true
        responses:
          '200':
            description: Workspace mis à jour
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
                    id:
                      type: integer
                    name:
                      type: string
                    status:
                      type: boolean
          '400':
            description: Requête invalide
          '404':
            description: Workspace non trouvé
          '401':
            description: Non autorisé
...
  /api/workspaces:
    get:
      tags:
        - Workspaces
      summary: Lister les workspaces de l'utilisateur connecté
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Liste des workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workspaces'

    post:
      tags:
        - Workspaces
      summary: Créer un nouveau workspace
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                status:
                  type: boolean
              required:
                - name
                - status
      responses:
        '201':
          description: Workspace créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspaces'
        '409':
          description: Workspace déjà existant
  /api/workspaces/{id}:
    get:
      tags:
        - Workspaces
      summary: Afficher un workspace
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Détails du workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspaces'

    delete:
      tags:
        - Workspaces
      summary: Supprimer un workspace
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Workspace supprimé
        '403':
          description: Accès refusé
        '400':
          description: Workspace non vide
  /api/workspaces/publics:
    get:
      tags:
        - Workspaces
      summary: Lister les workspaces publics non rejoints
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Liste des workspaces publics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workspaces'
  /api/workspaces/{id}/join:
    post:
      tags:
        - Workspaces
      summary: Rejoindre un workspace public
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Déjà membre
        '201':
          description: Ajouté avec succès
        '404':
          description: Workspace introuvable ou non public
        '500':
          description: Rôle introuvable